refact: made all saving/loading in one place
