+++ display/app/api/teams/info/route.ts
@@ -1,4 +1,4 @@
-import { getData } from "@/lib/server/dataImport";
+import { getData } from "@/lib/server/storage";
 import { NextRequest, NextResponse } from "next/server";
 
 export function GET(request: NextRequest) {
