+++ display/lib/server/breakHandler.js
@@ -1,5 +1,7 @@
 import fs from 'fs';
 import path from "path";
+
+import { load, save } from '../utils/localDataManager';
 import { getData } from '../utils/dataHelper';
 
 const breakInfo = await getData('/api/break_data/screens')
@@ -7,41 +9,8 @@ const breakInfo = await getData('/api/break_data/screens')
 const statePath = path.join(process.cwd(), "state/break.json");
 const stateDefaultPath = path.join(process.cwd(), "state/defaults/break.json")
 
-// Load from saved data
-function load() {
-    try {
-        const raw = fs.readFileSync(statePath, "utf8");
-        let obj = JSON.parse(raw);
-
-        return obj
-    } catch (err) {
-        console.error("Can't load break screen state, trying defaults");
-        return loadDefaults();
-    }
-}
-
-// Load from default state file
-function loadDefaults() {
-    try {
-        const raw = fs.readFileSync(stateDefaultPath, "utf8");
-        let obj = JSON.parse(raw);
-
-        return obj;
-    } catch (err) {
-        console.error("Can't load default break screen state.");
-        return {}
-    }
-}
-
-function save(state) {
-    try {
-        fs.writeFileSync(statePath, JSON.stringify(state, null, 2), "utf8");
-    } catch (err) {
-        console.error("Can't write voting state", error);
-    }
-}
-
-let data = load();
+let data = load(statePath);
+if (!data) data = loadDefaults(stateDefaultPath);
 
 /* --------------
     GETTERS
@@ -56,7 +25,7 @@ export function setBreakScreen(key) {
 
     if (breakData) {
         data.currentScreen = key;
-        save(data)
+        save(statePath, data);
         return true;
     }
 
@@ -66,6 +35,6 @@ export function setBreakScreen(key) {
 /* RESET */
 export function resetBreakScreen() {
     data = loadDefaults();
-    save(data);
+    save(statePath, data);
     return true;
 }
