+++ display/lib/server/votingHandler.js
@@ -3,6 +3,7 @@ import config from '@/config/general.json'
 import fs from 'fs';
 import path from "path";
 
+import { load, save } from '../utils/localDataManager';
 import { checkGame } from '../client/gameInfo';
 import { notify } from "@/lib/transmitter/listeners";
 
@@ -18,49 +19,14 @@ function setupSlotsAfterLoad(slots, slotsCount) {
     return slots;
 }
 
-// Load from saved data
-function load() {
-    try {
-        const raw = fs.readFileSync(statePath, "utf8");
-        let obj = JSON.parse(raw);
+// Setup
+let data = load(statePath);
+if (!data) data = loadDefaults(stateDefaultPath);
 
-        // Setting up
-        let slots = setupSlotsAfterLoad(obj.slots, config.voting.slots)
-        obj.slots = slots;
+// Setup voting slots
+let slots = setupSlotsAfterLoad(data.slots, config.voting.slots)
+data.slots = slots;
 
-        return obj
-    } catch (err) {
-        console.error("Can't load voting state, trying defaults");
-        return loadDefaults();
-    }
-}
-
-// Load from default state file
-function loadDefaults() {
-    try {
-        const raw = fs.readFileSync(stateDefaultPath, "utf8");
-        let obj = JSON.parse(raw);
-
-        // Setting up
-        let slots = setupSlotsAfterLoad(obj.slots, config.voting.slots)
-        obj.slots = slots;
-
-        return obj;
-    } catch (err) {
-        console.error("Can't load default voting state.");
-        return {}
-    }
-}
-
-function save(state) {
-    try {
-        fs.writeFileSync(statePath, JSON.stringify(state, null, 2), "utf8");
-    } catch (err) {
-        console.error("Can't write voting state", error);
-    }
-}
-
-let data = load();
 let currentSelectedSlot;
 let currentSelectedTimeout;
